优化角色分配菜单权限代码重复
