+++ system-service/src/main/java/com/ifox/platform/adminuser/request/role/RoleUpdateRequest.java
@@ -1,13 +1,13 @@
 package com.ifox.platform.adminuser.request.role;
 
 import com.ifox.platform.adminuser.dto.base.RoleBaseColumns;
+import com.ifox.platform.entity.sys.MenuPermissionEO;
 
+import java.util.ArrayList;
 import java.util.List;
 
 public class RoleUpdateRequest extends RoleBaseColumns {
     private String id;
-
-    private List<String> menuPermissions;
 
     public String getId() {
         return id;
@@ -17,19 +17,10 @@ public class RoleUpdateRequest extends RoleBaseColumns {
         this.id = id;
     }
 
-    public List<String> getMenuPermissions() {
-        return menuPermissions;
-    }
-
-    public void setMenuPermissions(List<String> menuPermissions) {
-        this.menuPermissions = menuPermissions;
-    }
-
     @Override
     public String toString() {
         return "RoleUpdateRequest{" +
             "id='" + id + '\'' +
-            ", menuPermissions=" + menuPermissions +
             "} " + super.toString();
     }
 }
